core::rt Restructure some modules
Put all uv code under rt::uv, as if it were in its own crate.
Pull local_sched out of rt::sched.
